# apps/docs/content/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/webhooks/{id}/secret/roll": {
      "post": {
        "operationId": "WebhookController_rollSigningSecret",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollSecretWebhookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Roll Webhook  Signing Secret",
        "tags": [
          "webhooks"
        ]
      }
    },
...
    "/accounts": {
      "get": {
        "operationId": "AccountController_getAccounts",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "example": "-state",
              "type": "string"
            }
          },
          {
            "name": "expand",
            "required": false,
            "in": "query",
            "description": "List of relations to be populated.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "List of fields to select.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filter options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "or",
            "required": false,
            "in": "query",
            "description": "Or options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "nor",
            "required": false,
            "in": "query",
            "description": "Nor options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "and",
            "required": false,
            "in": "query",
            "description": "And options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "projection",
            "required": false,
            "in": "query",
            "description": "Filter Projection",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "countTotal",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "dryRun",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get All Accounts",
        "tags": [
          "accounts"
        ]
      },
      "post": {
        "operationId": "AccountController_create",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "example": "-state",
              "type": "string"
            }
          },
          {
            "name": "expand",
            "required": false,
            "in": "query",
            "description": "List of relations to be populated.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "List of fields to select.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filter options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "or",
            "required": false,
            "in": "query",
            "description": "Or options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "nor",
            "required": false,
            "in": "query",
            "description": "Nor options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "and",
            "required": false,
            "in": "query",
            "description": "And options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "projection",
            "required": false,
            "in": "query",
            "description": "Filter Projection",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "countTotal",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "dryRun",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create Account",
        "tags": [
          "accounts"
        ]
      }
    },
    "/accounts/{id}": {
      "get": {
        "operationId": "AccountController_getOne",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "example": "-state",
              "type": "string"
            }
          },
          {
            "name": "expand",
            "required": false,
            "in": "query",
            "description": "List of relations to be populated.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "List of fields to select.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filter options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "or",
            "required": false,
            "in": "query",
            "description": "Or options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "nor",
            "required": false,
            "in": "query",
            "description": "Nor options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "and",
            "required": false,
            "in": "query",
            "description": "And options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "projection",
            "required": false,
            "in": "query",
            "description": "Filter Projection",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "countTotal",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "dryRun",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get One Account",
        "tags": [
          "accounts"
        ]
      }
    },
    "/accounts/{id}/balance/history": {
      "get": {
        "operationId": "AccountController_getBalanceHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "example": "-state",
              "type": "string"
            }
          },
          {
            "name": "expand",
            "required": false,
            "in": "query",
            "description": "List of relations to be populated.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "List of fields to select.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filter options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "or",
            "required": false,
            "in": "query",
            "description": "Or options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "nor",
            "required": false,
            "in": "query",
            "description": "Nor options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "and",
            "required": false,
            "in": "query",
            "description": "And options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "projection",
            "required": false,
            "in": "query",
            "description": "Filter Projection",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "countTotal",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "dryRun",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get Balance History",
        "tags": [
          "accounts"
        ]
      }
    },
    "/accounts/{id}/balance/history/csv": {
      "get": {
        "operationId": "AccountController_getBalanceHistoryCSV",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "example": "-state",
              "type": "string"
            }
          },
          {
            "name": "expand",
            "required": false,
            "in": "query",
            "description": "List of relations to be populated.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "List of fields to select.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filter options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "or",
            "required": false,
            "in": "query",
            "description": "Or options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "nor",
            "required": false,
            "in": "query",
            "description": "Nor options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "and",
            "required": false,
            "in": "query",
            "description": "And options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "projection",
            "required": false,
            "in": "query",
            "description": "Filter Projection",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "countTotal",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "dryRun",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Download Balance Logs in CSV",
        "tags": [
          "accounts"
        ]
      }
    },
    "/accounts/{id}/statement/pdf": {
      "get": {
        "operationId": "AccountController_getStatementPDF",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          {
            "name": "timezone",
            "required": false,
            "in": "query",
            "description": "Timezone",
            "schema": {
              "example": "Africa/Lagos",
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "description": "Start Date",
            "schema": {
              "example": "2021-01-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "description": "End Date",
            "schema": {
              "example": "2021-12-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Download Statement in PDF",
        "tags": [
          "accounts"
        ]
      }
    },
    "/accounts/{id}/balance": {
      "get": {
        "operationId": "AccountController_getBalance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get Balance",
        "tags": [
          "accounts"
        ]
      }
    },
    "/accounts/{id}/deposit-channels": {
      "post": {
        "operationId": "AccountController_addDepositChannel",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "example": "-state",
              "type": "string"
            }
          },
          {
            "name": "expand",
            "required": false,
            "in": "query",
            "description": "List of relations to be populated.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "List of fields to select.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filter options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "or",
            "required": false,
            "in": "query",
            "description": "Or options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "nor",
            "required": false,
            "in": "query",
            "description": "Nor options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "and",
            "required": false,
            "in": "query",
            "description": "And options",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "projection",
            "required": false,
            "in": "query",
            "description": "Filter Projection",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "countTotal",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "dryRun",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddDepositChannelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Add Deposit Channel",
        "tags": [
          "accounts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Propeller APIs",
    "description": "Financial Infrastructure for the China–Africa corridor",
    "version": "2024-05-01",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Object": {
        "type": "object",
        "properties": {}
      },
      "ObjectId": {
        "type": "object",
        "properties": {}
      },
...
      "RollSecretWebhookDto": {
        "type": "object",
        "properties": {
          "signingKey": {
            "type": "string"
          }
        }
      },
...
      "CreateAccountDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "settlementAccount": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "reference": {
            "type": "string",
            "description": "Account Reference"
          },
          "customer": {
            "type": "string"
          },
          "currency": {
            "type": "string",
            "example": "NGN",
            "enum": [
              "NGN",
              "USD"
            ]
          },
          "type": {
            "type": "string",
            "example": "main",
            "enum": [
              "main",
              "sub",
              "virtual"
            ]
          },
          "depositChannels": {
            "type": "string",
            "example": [
              "bank-account"
            ],
            "enum": [
              "bank-account"
            ]
          },
          "metadata": {
            "type": "object"
          }
        },
        "required": [
          "name",
          "currency",
          "type"
        ]
      },
      "AddDepositChannelDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "bank-account",
            "enum": [
              "bank-account"
            ]
          }
        },
        "required": [
          "type"
        ]
      },
...
    }
  }
}